// Crypto-app/src/interfaces/SingleCoinProps.ts
export interface SingleCoinProps {
    additional_notices: []
    asset_platform_id: string
    block_time_in_minutes: number
    categories: string[]
    coingecko_rank: number
    coingecko_score: number
    community_data: any
    community_score: number
    contract_address: string
    country_origin: string
    description: any
    detail_platforms: any
    developer_data: any
    developer_score: number
    genesis_date: null
    hashing_algorithm: null
    id: string
    image: any
    last_updated: string
    links: any
    liquidity_score: number
    localization: any
    market_cap_rank: number
    market_data: any
    name: string
    platforms: any
    public_interest_score: number
    public_interest_stats: any
    public_notice: null
    sentiment_votes_down_percentage: number
    sentiment_votes_up_percentage: number
    status_updates: []
    symbol: string
    tickers: any
}